requirements.txt
requests==2.26.0
python-dotenv==0.19.1
tqdm==4.62.3